working bls on 0.8.3 tests
