+++ packages/bls/package.json
@@ -28,7 +28,7 @@
     "pretest": "yarn check-types",
     "prepublishOnly": "yarn build",
     "test:unit": "nyc --cache-dir .nyc_output/.cache -r lcov -e .ts mocha --colors -r ./.babel-register 'test/unit/**/*.test.ts' && nyc report",
-    "test:spec": "mocha --colors -r ./.babel-register 'test/spec/**/*.test.ts'",
+    "test:spec": "mocha --colors -r ts-node/register 'test/spec/**/*.test.ts'",
     "test:spec-min": "yarn run test:spec",
     "test": "yarn test:unit && yarn test:spec",
     "coverage": "codecov -F bls",
@@ -64,7 +64,7 @@
     "codecov": "^3.1.0",
     "eslint": "^5.14.1",
     "js-yaml": "^3.13.1",
-    "mocha": "^5.2.0",
+    "mocha": "^6.2.0",
     "nyc": "^13.3.0",
     "sinon": "^7.2.7",
     "supertest": "^4.0.2",
