+++ packages/eth2.0-spec-test-util/package.json
@@ -44,6 +44,8 @@
     "bn.js": "^5.0.0",
     "camelcase": "^5.3.1",
     "deepmerge": "^4.0.0",
+    "chai": "^4.2.0",
+    "mocha": "^6.2.0",
     "js-yaml": "^3.13.1",
     "v8-profiler-next": "^1.1.1"
   },
@@ -53,9 +55,7 @@
     "@types/chai": "^4.1.7",
     "@types/deepmerge": "^2.2.0",
     "@types/js-yaml": "^3.12.1",
-    "@types/mocha": "^5.2.7",
-    "chai": "^4.2.0",
-    "mocha": "^6.2.0"
+    "@types/mocha": "^5.2.7"
   },
   "peerDependencies": {
     "chai": "^4.2.0",
